Admin Blur Theme
I added admin blur theme on login, with default dashboard working after
the login with mongodb. TODO, add bad credentials or server errors on
login, fix theme and validations on registration and forgot password,
and start with admin dashboards features like the profile and products
that are the basic ones for the minimun feasible product.